% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusPat}
\alias{clusPat}
\title{output from running clusterExpressionPattern from TradeSeq}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(clusPat, package = "SCMultiomeCNRSR")
}
\description{
Data generated after clustering the genes along the pseudotime for each lineage.
}
\keyword{datasets}
