% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster1.markers_RNA_snn_res.0.1}
\alias{cluster1.markers_RNA_snn_res.0.1}
\title{wilcox DE test results for clustering results}
\format{
An object of class \code{data.frame} with 9914 rows and 5 columns.
}
\usage{
data(cluster1.markers_RNA_snn_res.0.1, package = "SCMultiomeCNRSR")
}
\description{
Data generated after running FindMarkers from Seurat for cluster 1
after performing louvain clustering. Output in a SeuratObject.
}
\keyword{datasets}
