% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hippocampus_all_ages}
\alias{hippocampus_all_ages}
\title{mouse dentate gyrus hippocampus unprocessed scRNA-seq dataset}
\format{
An object of class \code{SingleCellExperiment} with 27932 rows and 24185 columns.
}
\source{
Gene Expression Omnibus with the accession code GSE104323.
}
\usage{
data(hippocampus_all_ages, package = "SCMultiomeCNRSR")
}
\description{
Data from Hochgerner et al. (2018) which contains the raw counts matrix and metadata in a #' SingleCellExperiment Object
}
\references{
Hochgerner, H., Zeisel, A., Lönnerberg, P. et al.
Conserved properties of dentate gyrus neurogenesis across postnatal
development revealed by single-cell RNA sequencing.
Nat Neurosci 21, 290–299 (2018). https://doi.org/10.1038/s41593-017-0056-2
}
\keyword{datasets}
