% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hippocampus_minimal}
\alias{hippocampus_minimal}
\title{subset of mouse dentate gyrus hippocampus processed scRNA-seq dataset}
\format{
An object of class \code{SingleCellExperiment} with 27932 rows and 5412 columns.
}
\usage{
data(hippocampus_minimal, package = "SCMultiomeCNRSR")
}
\description{
the mouse dentate gyrus hippocampus processed scRNA-seq dataset further
subsetted for cells in P0 and P5 stages showcasing cell types from two lineages
}
\keyword{datasets}
