% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hippocampus_processed}
\alias{hippocampus_processed}
\title{mouse dentate gyrus hippocampus processed scRNA-seq dataset}
\format{
An object of class \code{SingleCellExperiment} with 27932 rows and 18297 columns.
}
\usage{
data(hippocampus_processed, package = "SCMultiomeCNRSR")
}
\description{
Data from Hochgerner et al. (2018) which contains the raw counts matrix and metadata in a #' SingleCellExperiment Object after performing basic pre-processing
}
\keyword{datasets}
