% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scHOT_obj}
\alias{scHOT_obj}
\title{scHOT output to identify differentially correlated gene pairs.}
\format{
An object of class \code{scHOT} with 27932 rows and 2672 columns.
}
\usage{
data(scHOT_obj, package = "SCMultiomeCNRSR")
}
\description{
Data generated after running scHOT function to identify gene pairs that vary
in correlation across pseudotime.
}
\keyword{datasets}
