% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slingshot_umap}
\alias{slingshot_umap}
\title{output from fitting the NB-GAM model in TradeSeq}
\format{
An object of class \code{SingleCellExperiment} with 27932 rows and 5412 columns.
}
\usage{
data(slingshot_umap, package = "SCMultiomeCNRSR")
}
\description{
Data generated after running fitGAM to model gene expression changes as a
function of pseudotime. Output as a SlingShot Object.
}
\keyword{datasets}
